refactor(frontend): altering components
