+++ apps/frontend/src/components/ui/button.jsx
@@ -1,7 +1,7 @@
 import { Slot } from "@radix-ui/react-slot";
 import { cva } from "class-variance-authority";
 import PropTypes from "prop-types";
-import { cn } from "../../lib/utils";
+import { cn } from "../../utils/utils";
 
 const buttonVariants = cva(
   "inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-md text-sm font-medium transition-all disabled:pointer-events-none disabled:opacity-50 outline-none",
